scripts/requirements.txt
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.11.0
